fix issues with cn
